+++ ext/ssr_deno/crates/ssr_deno_dev_mode/src/dev_mode_builder.rs
@@ -128,6 +128,8 @@ pub fn build_dev_mode_worker(
             ..Default::default()
         }],
         startup_snapshot: None,
+        residual_lazy_js_sources: &[],
+        residual_lazy_esm_sources: &[],
         skip_op_registration: false,
         create_params,
         unsafely_ignore_certificate_errors: None,