+++ ext/ssr_deno/src/engine/builder.rs
@@ -120,6 +120,8 @@ pub fn build_worker(
             ..Default::default()
         }],
         startup_snapshot: None,
+        residual_lazy_js_sources: &[],
+        residual_lazy_esm_sources: &[],
         skip_op_registration: false,
         create_params,
         unsafely_ignore_certificate_errors: None,